requirements/base.txt
# Base dependencies for Flask API Template
# These are the core dependencies required for the application to run

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-RESTX==1.2.0
python-dotenv==1.0.0
marshmallow==3.20.1
psutil==5.9.5
gunicorn==21.2.0
